requirements.txt
flask==3.1.1
PyJWT
uvicorn==0.30.0
asgiref==3.9.0